decrease settings content and aggregate restapi methods
